LRAC-10551 Move BaseAnalyticsClientImpl to api package
