+++ modules/apps/analytics/analytics-message-sender-api/src/main/java/com/liferay/analytics/message/sender/client/BaseAnalyticsClientImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.analytics.message.sender.internal;
+package com.liferay.analytics.message.sender.client;
 
 import com.liferay.analytics.message.storage.service.AnalyticsMessageLocalService;
 import com.liferay.analytics.settings.configuration.AnalyticsConfiguration;
